Refactor report tables
