+++ src/scenario/index.ts
@@ -4,3 +4,4 @@ export * from './route-timetable'
 export * from './buffer-passes'
 export * from './stop-clusters'
 export * from './phases'
+export * from './reportTables'
